pyproject.toml
[project]
name = "cc-feishu-bridge"
version = "0.1.0"
description = "Claude Code Feishu Bridge"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.2.0",
    "lark-oapi>=1.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
cc-feishu-bridge = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
